+++ src/containers/NotFound.jsx
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import Helmet from 'react-helmet';
 
 export default class NotFound extends Component {
 
@@ -7,6 +8,7 @@ export default class NotFound extends Component {
   render() {
     return (
       <div>
+        <Helmet title="Not Found" />
         <h1>Not Found</h1>
       </div>
     );